requirements.txt
streamlit>=1.44
pdfplumber>=0.11
pymupdf>=1.25
pillow>=11.0
pytesseract>=0.3.13
qdrant-client>=1.12
requests>=2.32
python-multipart>=0.0.9
openpyxl>=3.1
